requirements.txt
fastapi==0.115.7
uvicorn==0.30.6
pymysql==1.1.1
python-dotenv==1.0.0
pydantic==2.9.2
